+++ front/app/components/admin/ImageCropper/ImageCropper.test.tsx
@@ -12,13 +12,23 @@ describe('ImageCropper', () => {
   it('renders image when there is an image', async () => {
     const image = (await getImage()) as UploadFile;
     render(
-      <ImageCropper image={image} onComplete={jest.fn()} onRemove={jest.fn()} />
+      <ImageCropper
+        image={image}
+        aspect={3 / 1}
+        onComplete={jest.fn()}
+        onRemove={jest.fn()}
+      />
     );
     expect(screen.getByRole('img')).toBeInTheDocument();
   });
   it('does not render when there is no image', () => {
     render(
-      <ImageCropper image={null} onComplete={jest.fn()} onRemove={jest.fn()} />
+      <ImageCropper
+        image={null}
+        onComplete={jest.fn()}
+        aspect={3 / 1}
+        onRemove={jest.fn()}
+      />
     );
     expect(screen.queryByRole('img')).not.toBeInTheDocument();
   });