+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
+## 2023-01-16
+
+### Added
+
+- [CL-2320] Image dimensions i2: folder & project banner images
+
 ## 2023-01-13 (2)
 
 ### Added
 
 - [CL-2214] Report builder MVP (behind feature flag)
+
+### Fixed
+
+- [CL-2597] Custom tags tab is once again visible in both projects and platform settings
 
 ## 2023-01-13